requirements.txt
python-socketio
aiohttp
requests
asyncio
